.history/requirements_20250910162653.txt
requests
PyYAML
pymongo
python-dateutil
python-dotenv
